feat: delete button fully working
