--- src/components/EditButton.vue
@@ -1,4 +1,0 @@
-<script setup></script>
-<template>
-  <button @click="editProduction">🛠️</button>
-</template>
